backend/requirements.txt
asgiref==3.7.2
attrs==23.2.0
Django==5.0.3
django-cors-headers==4.4.0
django-environ==0.11.2
django-extensions==3.2.3
django-rest-knox==5.0.1
djangorestframework==3.14.0
drf-spectacular==0.27.2
gunicorn==22.0.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
packaging==24.0
psycopg==3.1.18
psycopg2==2.9.9
pydot==2.0.0
pyparsing==3.1.2
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.19.1
sqlparse==0.4.4
typing_extensions==4.10.0
uritemplate==4.1.1
zlib-ng==0.4.3
